remove the unnecessary code after discuss
