+++ src/main/kotlin/dependencyinjection/projectModule.kt
@@ -4,7 +4,7 @@ import org.example.data.CsvMealParser
 import org.example.data.CsvMealReader
 import org.example.data.CsvRepositoryImpl
 import org.example.logic.CsvRepository
-import org.example.logic.EasyFoodSuggestion
+import org.example.logic.EasyFoodSuggestionUseCase
 import org.koin.dsl.module
 
 val projectModule = module {
@@ -13,6 +13,6 @@ val projectModule = module {
     single { CsvMealParser() }
 
     single<CsvRepository> { CsvRepositoryImpl(get(), get()) }
-    single { EasyFoodSuggestion(get()) } // Add this line
 
+    single { EasyFoodSuggestionUseCase(get()) }
 }
