--- src/main/kotlin/model/FormattedMeal.kt
@@ -1,10 +1,0 @@
-package org.example.model
-
-data class FormattedMeal(
-    val name: String,
-    val preparationTime: String,
-    val ingredients: String,
-    val steps: String,
-    val nutritionSummary: String,
-    val tags: String
-)
